==> 2_user_label_assessment.Rmd <==
---
title: "Assessment of User Labels"
author: "Bob Horton"
date: "November 22, 2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, cache=FALSE, message=FALSE)
library(knitr)
rxOptions(reportProgress=0)
```

## Performance of models trained on user-labeled data

Here we add the labels provided by each user to the initial training set and assess the effects of this newly labelled data on model performance.

```{r parameters}

### Libraries ###
library(dplyr)
library(pROC)
library(ggplot2)
library(knitr)

### Meta-hyperparameters ###
set.seed(1)  ###

L1_PENALTY <- 1e-2
L2_PENALTY <- 1e-2

INITIAL_EXAMPLES_PER_CLASS <- 6  # cases from the labelled dataset used to train the initial model

# This order determines the order of factor levels
KNOT_CLASSES <- setNames(nm=c("sound_knot", "dry_knot", "encased_knot"))

LABELLED_FEATURIZED_DATA <- "data/labelled_knots_featurized_resnet18.Rds"
UNLABELLED_FEATURIZED_DATA <- "data/unlabelled_knots_featurized_resnet18.Rds"

unlabelled_knot_data_df <- readRDS(UNLABELLED_FEATURIZED_DATA)
# labelled_knot_data_df <- readRDS(LABELLED_FEATURIZED_DATA)

TEST_SET <- readRDS("data/test_set.Rds")
INITIAL_TRAINING_SET <- readRDS("data/initial_training_set.Rds")

inputs <- grep("^Feature", names(INITIAL_TRAINING_SET), value=TRUE)
outcome <- "knot_class"
FORM <- formula(paste(outcome, paste(inputs, collapse="+"), sep="~"))

STANDARD_LABEL_SET <- "f6f5457"

```


# User labels

These are the labels collected from workshop participants.

```{r user_label_data}

source("secret.R")

conn_info <- list(
  Driver="{ODBC Driver 13 for SQL Server}",
  Server="tcp:jichangserver.database.windows.net,1433",
  Database="woodknots",
  Uid=secret$Uid,
  Pwd=secret$Pwd,
  Encrypt="yes",
  TrustServerCertificate="no",
  `Connection Timeout`="30;"
)

conn_str <- paste(paste(names(conn_info), conn_info, sep="="), collapse=";")


library(RODBC)
dbhandle <- odbcDriverConnect(conn_str)
sql <- function(q) sqlQuery(dbhandle, q) # same syntax as sqldf

# sql("SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TABLE'")

# sql("SELECT top 5 * FROM UserKnotLabel")

  
(user_knot_labels <- sql("SELECT
                              substring(userid, 1, 7) as user_id,
                              knotname,
                              label_value + '_knot' as knot_class,
                              ukl.labeltime labeltime
                          FROM UserKnotLabel ukl JOIN Label ON ukl.labelid = Label.label_id
                            WHERE ukl.labeltime > DATEADD(HH,-2, GETDATE())
                            ORDER BY user_id, ukl.labeltime"))

# To select cases from a particular time range:

# (user_knot_labels <- sql("SELECT 
#                               substring(userid, 1, 7) as user_id, 
#                               knotname, 
#                               label_value + '_knot' as knot_class, 
#                               ukl.labeltime labeltime
#                           FROM UserKnotLabel ukl JOIN Label ON ukl.labelid = Label.label_id
#                             WHERE ukl.labeltime > CONVERT(datetime, '2017-11-22 04:45:00', 120)
#                               AND ukl.labeltime < CONVERT(datetime, '2017-11-22 06:00:00', 120)
#                             ORDER BY user_id, ukl.labeltime"))

```

```{r source_library_code}

source("woodknots_active_learning_lib.R")

```


# User Results

These are the performance scores for models built with the initial training set plus the new cases as labelled by each user.

```{r evaluate_user_labelled_datasets}

user_id_vec <- setNames(nm=levels(user_knot_labels$user_id))

user_results <- lapply(user_id_vec, function(userID){
  user_labels <- user_knot_labels %>% 
                    filter(user_id == userID) %>% 
                    group_by(knotname) %>% 
                    filter(labeltime==max(labeltime)) %>% # only keep the last label they chose
                    as.data.frame
  
  row.names(user_labels) <- user_labels$knotname

  user_cases <- unlabelled_knot_data_df %>%
                    filter(path %in% user_labels$knotname)

  user_cases$knot_class <- user_labels[user_cases$path, "knot_class"]

  user_training_set <- rbind(INITIAL_TRAINING_SET, user_cases)

  fit_and_evaluate_model(user_training_set)
}) %>% sapply(function(ures) ures$performance) %>% t %>% as.data.frame

users_ordered <- row.names(user_results)[order(user_results$accuracy, decreasing = FALSE)]

if (STANDARD_LABEL_SET %in% users_ordered){
  users_ordered <- c(setdiff(users_ordered, STANDARD_LABEL_SET), STANDARD_LABEL_SET)
}
  
user_results$user_id <- factor(row.names(user_results), levels=users_ordered)

user_results <- user_results[users_ordered,]

user_results %>% 
  mutate(is_standard = user_id==STANDARD_LABEL_SET) %>%
  ggplot(aes(x=user_id, y=accuracy, col=is_standard)) + 
  geom_col() + 
  coord_flip() + 
  ggtitle("User results ranked by accuracy")

knitr::kable(user_results[rev(users_ordered),])

```
